+++ plugins/pragmatach-json/src/test/java/test/com/khubla/pragmatach/plugin/json/pojo/MapPOJO.java
@@ -1,0 +1,18 @@
+package test.com.khubla.pragmatach.plugin.json.pojo;
+
+import java.util.Map;
+
+/**
+ * @author tome
+ */
+public class MapPOJO {
+   private Map<String, String> stringMap;
+
+   public Map<String, String> getStringMap() {
+      return stringMap;
+   }
+
+   public void setStringMap(Map<String, String> stringMap) {
+      this.stringMap = stringMap;
+   }
+}