+++ pragmatach-framework/src/main/java/com/khubla/pragmatach/framework/application/Application.java
@@ -1,7 +1,10 @@
 package com.khubla.pragmatach.framework.application;
 
 import com.khubla.pragmatach.framework.api.Configuration;
+import com.khubla.pragmatach.framework.api.PragmatachException;
+import com.khubla.pragmatach.framework.controller.Controllers;
 import com.khubla.pragmatach.framework.lifecycle.LifecycleListeners;
+import com.khubla.pragmatach.framework.router.PragmatachRoutes;
 
 /**
  * @author tome
@@ -20,8 +23,22 @@ public class Application {
       return configuration;
    }
 
+   /**
+    * get all known controllers
+    */
+   public static Controllers getControllers() {
+      return Controllers.getInstance();
+   }
+
    public static LifecycleListeners getLifecyclelisteners() {
       return lifecycleListeners;
+   }
+
+   /**
+    * get all known routes
+    */
+   public static PragmatachRoutes getRoutes() throws PragmatachException {
+      return PragmatachRoutes.getInstance();
    }
 
    public static void setConfiguration(Configuration configuration) {