+++ pragmatach-framework/src/main/java/com/khubla/pragmatach/framework/controller/BeanBoundController.java
@@ -1,12 +1,10 @@
 package com.khubla.pragmatach.framework.controller;
 
-import java.util.Map;
-
 import com.khubla.pragmatach.framework.api.PragmatachException;
 
 /**
  * @author tome
  */
 public interface BeanBoundController extends PragmatachController {
-   Map<String, String> getPostFieldValues() throws PragmatachException;
+   void populateController() throws PragmatachException;
 }