+++ pragmatach-framework/src/main/java/com/khubla/pragmatach/framework/router/PragmatachRoutes.java
@@ -51,10 +51,26 @@ public class PragmatachRoutes {
       readRoutes();
    }
 
+   /**
+    * get all known routes for all HTTP methods
+    */
+   public List<PragmatachRoute> getAllRoutes() {
+      final List<PragmatachRoute> ret = new ArrayList<PragmatachRoute>();
+      ret.addAll(GETRoutes);
+      ret.addAll(POSTRoutes);
+      return ret;
+   }
+
+   /**
+    * get all GET routes
+    */
    public List<PragmatachRoute> getGETRoutes() {
       return GETRoutes;
    }
 
+   /**
+    * get all POST routes
+    */
    public List<PragmatachRoute> getPOSTRoutes() {
       return POSTRoutes;
    }